refactor/ Entity Update Done.
