+++ src/main/kotlin/com/wwdt/WwdtApplication.kt
@@ -2,8 +2,10 @@ package com.wwdt
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing
 
 @SpringBootApplication
+@EnableJpaAuditing
 class WwdtApplication
 
 fun main(args: Array<String>) {
