--- src/main/kotlin/com/wwdt/domain/entity_deprecated/IssueTeam.kt
@@ -1,37 +1,0 @@
-package com.wwdt.domain.entity_deprecated
-
-import com.wwdt.domain.entity_deprecated.common.BaseEntity
-import jakarta.persistence.*
-
-@Entity
-@Table(name = "t_issue_team")
-class IssueTeam(
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "issue_id")
-    val issue: Issue,
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "team_seq")
-    val team: Team,
-
-    @Id
-    @GeneratedValue
-    val seq: Long = 0L
-): BaseEntity(){
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other !is IssueTeam) return false
-
-        if (seq != other.seq) return false
-
-        return true
-    }
-
-    override fun hashCode(): Int {
-        return seq.hashCode()
-    }
-
-    override fun toString(): String {
-        return "IssueTeam(issue=$issue, team=$team, seq=$seq)"
-    }
-}
