--- src/main/kotlin/com/wwdt/domain/entity_deprecated/enums/RoleGrant.kt
@@ -1,7 +1,0 @@
-package com.wwdt.domain.entity_deprecated.enums
-
-enum class RoleGrant {
-    ROLE_ADMIN,
-    ROLE_TEAM_LEADER,
-    ROLE_TEAM_MEMBER
-}
