+++ src/main/kotlin/com/wwdt/domain/model/Silo.kt
@@ -1,5 +1,6 @@
 package com.wwdt.domain.model
 
+import com.wwdt.shared_kernal.model.BaseEntity
 import jakarta.persistence.*
