+++ src/main/kotlin/com/wwdt/domain/model/enums/IssuePriority.kt
@@ -1,4 +1,4 @@
-package com.wwdt.domain.entity_deprecated.enums
+package com.wwdt.domain.model.enums
 
 enum class IssuePriority {
     LOWEST,
